requirements.txt
# 基础依赖
langchain>=0.1.0
langgraph>=0.0.15
pydantic>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
pytest>=7.3.1
# 模型服务
openai>=1.0.0
